+++ packages/kolibri-tools/lib/extract$trs.js
